packages/ml/requirements.txt
# FastAPI and server
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.9.0
pydantic-settings>=2.2.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.25
# asyncpg==0.29.0  # Only needed for PostgreSQL
aiosqlite>=0.19.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# ML Core
scikit-learn>=1.4.0
xgboost>=2.0.3
lightgbm>=4.2.0

# Time series forecasting (optional - not required for basic functionality)
# prophet>=1.1.5
# statsmodels>=0.14.1
# darts==0.27.0  # Optional: uncomment for advanced time series

# Recommendation systems (optional - not required for basic functionality)
# implicit>=0.7.2  # Not compatible with Python 3.13
# lightfm>=1.17  # Not compatible with Python 3.13
scipy>=1.11.4

# NLP and embeddings
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
transformers>=4.37.0

# Model serialization and utilities
joblib>=1.3.2
cloudpickle>=3.0.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0

# Development
python-dotenv>=1.0.0
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
